Fixing writer issue.

+++ src/main.rs
@@ -38,7 +38,7 @@ fn main() {
             commands::report(load_configuration(config_path.to_string()));
         }
         ("start", Some(_submatches)) => {
-            let sleep_time = 100000; // millis
+            let sleep_time = 100000; // millis, to be moved to config blob.
             Command::new("./is-up-daemon")
                 .arg(sleep_time.to_string())
                 .arg("is-up")
@@ -48,11 +48,18 @@ fn main() {
 
         ("stop", Some(_submatches)) => {
             let pid = fs::read_to_string("/tmp/is-up.pid").expect("PID file (/tmp/is-up.pid) could not be read. It is likely the daemon is not running.");
-            // TODO: validate format of PID?
+            let pid = pid.parse::<u64>().expect(
+                "PID was non-numeric. Cowardly refusing to try to kill a non-numeric process.",
+            );
+
+            if pid <= 1 {
+                panic!("Invalid PID in /tmp/is-up.pid.");
+            }
+
             Command::new("kill")
-                .arg(pid)
+                .arg(pid.to_string())
                 .output()
-                .expect("failed to kill");
+                .expect("Failed to kill process");
             fs::remove_file("/tmp/is-up.pid").expect(
                 "Killed binary but could not remove /tmp/is-up.pid. Please remove manually.",
             );
